+++ src/providers/bedrock/types.ts
@@ -1,5 +1,4 @@
-
 export enum BEDROCK_LLAMA_STOP_REASON {
   stop = 'stop',
-  length = 'length'
+  length = 'length',
 }
